data/state/prop reorganization
