+++ portfolio-v1/src/data/ExperienceData.js
@@ -1,31 +1,33 @@
-import persado_logo_white from "../../assets/persado_logo_white.png";
-import persado_logo_black from "../../assets/persado-logo-black.png";
-import GA_Logo2 from "../../assets/GA_Logo2.png";
-import carhartt_wip_text_logo_white from "../../assets/carhartt_wip_text_logo-white.png";
-import carhartt_wip_text_logo_black from "../../assets/carhartt_wip_text_logo.png";
-import Unis from "../../assets/Unis.png";
-import adobephotoshop from "../../assets/Skills/adobephotoshop.png";
-import asana from "../../assets/Skills/asana.png";
-import bootstrap from "../../assets/Skills/bootstrap.png";
-import css3 from "../../assets/Skills/css3-alt.png";
-import django from "../../assets/Skills/django.png";
-import html5 from "../../assets/Skills/html5.png";
-import git from "../../assets/Skills/icons8-git.png";
-import excel from "../../assets/Skills/icons8-microsoft-excel.png";
-import mongodb from "../../assets/Skills/icons8-mongodb.png";
-import python from "../../assets/Skills/icons8-python.png";
-import reactnative from "../../assets/Skills/icons8-react-native.png";
-import visualstudio from "../../assets/Skills/icons8-visual-studio-code-2019.png";
-import js from "../../assets/Skills/js.png";
-import mailchimp from "../../assets/Skills/mailchimp copy 2.png";
-import node from "../../assets/Skills/node-js.png";
-import postgresql from "../../assets/Skills/postgresql.png";
-import react from "../../assets/Skills/react.png";
-import semanticuireact from "../../assets/Skills/semanticuireact.png";
-import shopify from "../../assets/Skills/shopify.png";
-import sketch from "../../assets/Skills/sketch.png";
+import persado_logo_white from "../assets/persado_logo_white.png";
+import persado_logo_black from "../assets/persado-logo-black.png";
+import GA_Logo2 from "../assets/GA_Logo2.png";
+import carhartt_wip_text_logo_white from "../assets/carhartt_wip_text_logo-white.png";
+import carhartt_wip_text_logo_black from "../assets/carhartt_wip_text_logo.png";
+import Unis from "../assets/Unis.png";
+import adobephotoshop from "../assets/Skills/adobephotoshop.png";
+import asana from "../assets/Skills/asana.png";
+import bootstrap from "../assets/Skills/bootstrap.png";
+import css3 from "../assets/Skills/css3-alt.png";
+import django from "../assets/Skills/django.png";
+import html5 from "../assets/Skills/html5.png";
+import git from "../assets/Skills/icons8-git.png";
+import excel from "../assets/Skills/icons8-microsoft-excel.png";
+import mongodb from "../assets/Skills/icons8-mongodb.png";
+import python from "../assets/Skills/icons8-python.png";
+import reactnative from "../assets/Skills/icons8-react-native.png";
+import visualstudio from "../assets/Skills/icons8-visual-studio-code-2019.png";
+import js from "../assets/Skills/js.png";
+import mailchimp from "../assets/Skills/mailchimp copy 2.png";
+import node from "../assets/Skills/node-js.png";
+import postgresql from "../assets/Skills/postgresql.png";
+import react from "../assets/Skills/react.png";
+import semanticuireact from "../assets/Skills/semanticuireact.png";
+import shopify from "../assets/Skills/shopify.png";
+import sketch from "../assets/Skills/sketch.png";
 
-export const ExperienceData = [
+
+
+export const Jobs = [
   {
     company: "Persado",
     title: "Front End Developer",
